+++ sources/elementsmover.cpp
@@ -17,6 +17,7 @@
 */
 #include "elementsmover.h"
 #include "qetproject.h"
+#include "autobreakconductor.h"
 #include "conductorautonumerotation.h"
 #include "diagram.h"
 #include "qetgraphicsitem/conductor.h"
@@ -176,6 +177,16 @@ void ElementsMover::endMovement()
 	if (!m_current_movement.isNull()) {
 		QUndoCommand *quc{new MoveGraphicsItemCommand(m_diagram, m_moved_content, m_current_movement, undo_object)};
 		undo_object->setText(quc->text());
+	}
+
+		//Auto-break conductors: for each moved element, break any conductor
+		//whose path passes through a terminal dock point, and reconnect through
+		//the element.  The element is already at its final position on screen.
+	if (m_diagram->project()->autoBreakConductor())
+	{
+		for (Element *e : m_moved_content.m_elements) {
+			autoBreakConductors(m_diagram, e, undo_object);
+		}
 	}
 
 		//There is only one element moved, and project authorize auto conductor,
